# Ensure all runtime_modules loaded before mapping their paths