@josevalim 2022-12-08
@naserca unfortunately I had to revert this because we cannot assume we can load all modules. The good news is that I believe we have a simpler fix which I will publish soon for you to try out. :)