@josevalim 2022-12-08
I have pushed a fix here: https://github.com/elixir-lang/elixir/commit/8f867100bb9d3bff1155907d9b355f64e85fc96d - can you please let me know how it goes? Thank you!