@naserca 2022-12-08
Not unfortunate at all! I didn't really expect for my naive one-liner to be the best approach :) The good news is yes:

I've cherry-picked your commit locally into our current version `v1.14.2` (adjusting some given the changes in `main`) and seeing the same general super-exciting effects on performance during a typical recompilation in our application:

Before:
```bash
mix compile  86.53s user 42.16s system 140% cpu 1:31.87 total
```

After:
```bash
mix compile  25.17s user 5.17s system 241% cpu 12.567 total
```

Huge thanks for taking a look! 💚 💙 💜 💛 ❤️