@josevalim 2022-12-08
Since you patched v1.14 already, could you send a PR backporting it to v1.14? :)